# openapi3-v1.yaml
openapi: 3.0.1
info:
  title: Job Manager Service - API v1
  description: Job Manager Service API version 1
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /v1/jobs:
    get:
      operationId: findJobs
      parameters:
        - $ref: '#/components/parameters/jobNameQueryParam'
        - $ref: '#/components/parameters/fromDate'
        - $ref: '#/components/parameters/endDate'
        - $ref: '#/components/parameters/priority'
        - $ref: '#/components/parameters/includeStages'
      summary: Retrieve jobs matching criteria
      description: |
        Filter jobs by name, date range, priority. Optional stage inclusion.
        Returns empty array if no matches.
      tags:
        - jobs
      responses:
        '200':
          description: Successfully retrieved matching jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/job'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
    post:
      operationId: createJob
      summary: Create job
      description: |
        Create job with config and metadata. Initial status: PENDING.
      tags:
        - jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createJobPayload'
            examples:
              simple_job:
                summary: Simple job with minimal configuration
                description: Basic job creation with default priority
                value:
                  name: data_processing
                  data:
                    input_path: /data/raw/batch_001
                    output_path: /data/processed/batch_001
                    format: parquet
              complex_job:
                summary: Complex job with full configuration
                description: Job with high priority and custom metadata
                value:
                  name: critical_analysis
                  data:
                    input_sources:
                      - /data/source1/*.csv
                      - /data/source2/*.json
                    transformations:
                      - type: filter
                        conditions:
                          - status = 'active'
                          - created_date > '2024-01-01'
                      - type: aggregate
                        group_by:
                          - region
                          - category
                        metrics:
                          - sum(revenue)
                          - count(transactions)
                    output_config:
                      format: delta
                      partitions:
                        - year
                        - month
                      compression: snappy
                  priority: VERY_HIGH
                  userMetadata:
                    department: analytics
                    owner: data-team@company.com
                    deadline: '2025-08-01T00:00:00Z'
                    cost_center: CC-12345
              job_with_tracing:
                summary: Job with explicit tracing context
                description: Job creation with provided distributed tracing context
                value:
                  name: traced_processing
                  data:
                    input_path: /data/traced/batch_001
                    output_path: /data/output/batch_001
                  traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
                  tracestate: rojo=00f067aa0ba902b7,congo=t61rcWkgMzE
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/job'
              examples:
                simple_job_response:
                  summary: Response for simple job creation
                  value:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    name: data_processing
                    status: PENDING
                    percentage: 0
                    priority: MEDIUM
                    creationTime: '2025-07-27T10:30:00Z'
                    updateTime: '2025-07-27T10:30:00Z'
                    data:
                      input_path: /data/raw/batch_001
                      output_path: /data/processed/batch_001
                      format: parquet
                    userMetadata: {}
                    traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
                complex_job_response:
                  summary: Response for complex job creation
                  value:
                    id: 660f9511-f3ac-52e5-b827-557766551111
                    name: critical_analysis
                    status: PENDING
                    percentage: 0
                    priority: VERY_HIGH
                    creationTime: '2025-07-27T10:30:00Z'
                    updateTime: '2025-07-27T10:30:00Z'
                    data:
                      input_sources:
                        - /data/source1/*.csv
                        - /data/source2/*.json
                      transformations:
                        - type: filter
                          conditions:
                            - status = 'active'
                            - created_date > '2024-01-01'
                        - type: aggregate
                          group_by:
                            - region
                            - category
                          metrics:
                            - sum(revenue)
                            - count(transactions)
                      output_config:
                        format: delta
                        partitions:
                          - year
                          - month
                        compression: snappy
                    userMetadata:
                      department: analytics
                      owner: data-team@company.com
                      deadline: '2025-08-01T00:00:00Z'
                      cost_center: CC-12345
                    traceparent: 00-660f9511f3ac52e5b827557766551111-22b378cc5ca902b7-01
                    tracestate: analytics=dept123,priority=very_high
        '400':
          description: Invalid request, could not create job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
              examples:
                validation_error:
                  summary: Validation error
                  description: Request failed due to missing required fields
                  value:
                    message: >-
                      Validation failed: 'name' is required and must be a
                      non-empty string. 'data' is required and must be an
                      object.
                    code: VALIDATION_ERROR
                invalid_priority:
                  summary: Invalid priority value
                  value:
                    message: >-
                      Invalid priority value 'SUPER_HIGH'. Valid values are:
                      VERY_HIGH, HIGH, MEDIUM, LOW, VERY_LOW.
                    code: VALIDATION_ERROR
                malformed_data:
                  summary: Malformed request data
                  value:
                    message: >-
                      Invalid JSON format in request body. Expected object but
                      received array at line 1, column 1.
                    code: VALIDATION_ERROR
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/jobs/{jobId}:
    parameters:
      - $ref: '#/components/parameters/jobId'
    get:
      operationId: getJobById
      parameters:
        - $ref: '#/components/parameters/includeStages'
      summary: Retrieve a specific job by its ID
      description: >
        Fetches detailed information about a job using its unique identifier.

        Includes job configuration, status, metadata, and completion percentage.


        Optional inclusion of related stage data via the should_return_stages
        parameter,

        which allows clients to retrieve the complete job hierarchy in a single
        request.
      tags:
        - jobs
      responses:
        '200':
          description: Job data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/job'
              examples:
                job_with_stages:
                  summary: Job with complete stage and task hierarchy
                  description: Full job details when should_return_stages=true
                  value:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    name: ml_pipeline
                    status: IN_PROGRESS
                    percentage: 45
                    priority: HIGH
                    creationTime: '2025-07-27T08:00:00Z'
                    updateTime: '2025-07-27T10:30:00Z'
                    traceparent: 00-660f9511f3ac52e5b827557766551111-22b378cc5ca902b7-01
                    data:
                      model_type: random_forest
                      training_data: /data/ml/training_set.parquet
                      features:
                        - age
                        - income
                        - location
                        - purchase_history
                      target: churn_probability
                      hyperparameters:
                        n_estimators: 100
                        max_depth: 10
                        min_samples_split: 5
                    userMetadata:
                      experiment_id: EXP-2025-001
                      researcher: ml-team@company.com
                      budget: $5000
                    stages:
                      - id: dd301d64-69b2-4b49-ba26-fb2a87fa0150
                        type: data_preparation
                        status: COMPLETED
                        percentage: 100
                        jobId: 550e8400-e29b-41d4-a716-446655440000
                        order: 1
                        data:
                          validation_split: 0.2
                          preprocessing_steps:
                            - normalize
                            - encode_categorical
                        userMetadata: {}
                        summary:
                          pending: 0
                          inProgress: 0
                          completed: 5
                          failed: 0
                          created: 0
                          retried: 0
                          total: 5
                        traceparent: >-
                          00-550e8400e29b41d4a716446655440000-22b378cc5ca902b7-01
                      - id: 2d9b2481-ec1c-4efd-9350-f33bd79d51eb
                        type: model_training
                        status: IN_PROGRESS
                        percentage: 60
                        jobId: 550e8400-e29b-41d4-a716-446655440000
                        order: 2
                        data:
                          cross_validation_folds: 5
                          early_stopping: true
                        userMetadata: {}
                        summary:
                          pending: 1
                          inProgress: 2
                          completed: 2
                          failed: 0
                          created: 0
                          retried: 0
                          total: 5
                        traceparent: >-
                          00-550e8400e29b41d4a716446655440000-22b378cc5ca902b7-01
                      - id: 637c05f3-e0b0-4c26-a791-89308c58e316
                        type: model_evaluation
                        status: WAITING
                        percentage: 0
                        jobId: 550e8400-e29b-41d4-a716-446655440000
                        order: 3
                        data:
                          metrics:
                            - accuracy
                            - precision
                            - recall
                            - f1_score
                          validation_dataset: /data/ml/validation_set.parquet
                        userMetadata: {}
                        summary:
                          pending: 3
                          inProgress: 0
                          completed: 0
                          failed: 0
                          created: 0
                          retried: 0
                          total: 3
                        traceparent: >-
                          00-550e8400e29b41d4a716446655440000-22b378cc5ca902b7-01
                simple_job:
                  summary: Simple job without stages
                  description: Basic job details when should_return_stages=false
                  value:
                    id: 660f9511-f3ac-52e5-b827-557766551111
                    name: data_backup
                    status: COMPLETED
                    percentage: 100
                    priority: LOW
                    creationTime: '2025-07-26T14:00:00Z'
                    updateTime: '2025-07-26T15:30:00Z'
                    traceparent: 00-660f9511f3ac52e5b827557766551111-22b378cc5ca902b7-01
                    data:
                      source_path: /data/production
                      destination: s3://backup-bucket/2025-07-26
                      compression: gzip
                      encryption: true
                    userMetadata:
                      backup_type: daily
                      retention_days: 30
        '400':
          description: Invalid request, could not get job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
    delete:
      operationId: deleteJob
      summary: Delete a job and all its associated resources (stages, tasks)
      description: >
        Permanently removes a job and all its associated stages and tasks from
        the system.

        This operation cascades to delete all child resources and cannot be
        undone.


        The job must exist in the system for this operation to succeed.

        Jobs can only be deleted when they are in a finite state (COMPLETED,
        FAILED, or ABORTED).

        Attempting to delete a job in any other state will result in a 400
        error.


        Returns a success message with code JOB_DELETED_SUCCESSFULLY when
        completed.
      tags:
        - jobs
      responses:
        '200':
          description: Job deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: JOB_DELETED_SUCCESSFULLY
        '400':
          description: Invalid parameters or validation error
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/baseErrorResponse'
                  - type: object
                    description: >
                      Error response structure for bad request errors, providing
                      a consistent

                      format for common-readable error codes and human-readable
                      messages.
                    required:
                      - code
                    properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
                          - JOB_NOT_IN_FINITE_STATE
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/jobs/{jobId}/user-metadata:
    patch:
      operationId: updateUserMetadata
      parameters:
        - $ref: '#/components/parameters/jobId'
      summary: Update job's custom metadata
      description: >
        Updates the user-defined metadata object for a specific job.

        This endpoint allows clients to attach or modify arbitrary data related
        to a job

        without affecting the job's core properties or execution status.


        User metadata is useful for storing application-specific context,
        tracking information,

        or any custom data needed by client applications.
      tags:
        - jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userMetadata'
      responses:
        '200':
          description: User metadata successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: JOB_MODIFIED_SUCCESSFULLY
        '400':
          description: Invalid metadata format or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/jobs/{jobId}/priority:
    parameters:
      - $ref: '#/components/parameters/jobId'
    patch:
      operationId: updateJobPriority
      summary: Modify job's priority level
      description: >
        Updates the priority level for a specific job.

        This affects how the job is scheduled relative to other jobs in the
        system.


        Higher priority jobs will be processed before lower priority ones when
        resources

        are constrained. Priority changes take effect immediately and apply to
        all 

        pending tasks associated with the job.
      tags:
        - jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                priority:
                  $ref: '#/components/schemas/priority'
              required:
                - priority
      responses:
        '200':
          description: Job priority successfully changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: JOB_MODIFIED_SUCCESSFULLY
        '204':
          description: No change made - requested priority equals current priority
          headers:
            Reason:
              description: Won't change priority if equal to current
              schema:
                type: string
          content: {}
        '400':
          description: Invalid priority value or other request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/jobs/{jobId}/status:
    parameters:
      - $ref: '#/components/parameters/jobId'
    put:
      operationId: updateStatus
      summary: Update job status for workflow control
      description: >
        Updates the operational status of a job to control its workflow
        execution. This endpoint 

        supports user-initiated operations including PENDING (resume), ABORTED
        (cancel), and 

        PAUSED (suspend), which may cascade changes to related stages and tasks.


        Internal status transitions (such as IN_PROGRESS, COMPLETED, FAILED,
        CREATED) are managed 

        automatically by the system based on stage and task completion, and
        cannot be set through 

        this endpoint.


        Status changes follow a state machine that enforces valid transitions.
        When a job's status 

        is changed, the system will automatically update timestamps and
        completion percentages as 

        appropriate.
      tags:
        - jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/jobOperationStatus'
              required:
                - status
      responses:
        '200':
          description: Job status successfully changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: JOB_MODIFIED_SUCCESSFULLY
        '400':
          description: Invalid status or illegal state transition
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/baseErrorResponse'
                  - type: object
                    description: >
                      Error response structure for bad request errors, providing
                      a consistent

                      format for common-readable error codes and human-readable
                      messages.
                    properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
                          - ILLEGAL_JOB_STATUS_TRANSITION
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
...
security: []
components:
  parameters:
    jobId:
      in: path
      name: jobId
      description: Unique identifier for the job
      required: true
      schema:
        $ref: '#/components/schemas/jobId'
...
    jobNameQueryParam:
      in: query
      name: job_name
      description: Filter jobs by their name/type
      required: false
      schema:
        $ref: '#/components/schemas/jobName'
    priority:
      in: query
      name: priority
      description: Filter jobs by their priority level
      required: false
      schema:
        $ref: '#/components/schemas/priority'
    fromDate:
      in: query
      name: from_date
      description: Filter results by update time, starting from this date/time
      required: false
      schema:
        type: string
        format: date-time
    endDate:
      in: query
      name: end_date
      description: Filter results by update time, ending at this date/time
      required: false
      schema:
        type: string
        format: date-time
    includeStages:
      in: query
      name: should_return_stages
      description: When true, includes stage data in the response
      required: false
      schema:
        $ref: '#/components/schemas/returnStage'
...
  schemas:
...
    priority:
      type: string
      enum:
        - VERY_HIGH
        - HIGH
        - MEDIUM
        - LOW
        - VERY_LOW
      example: LOW
      description: |
        Job priority for processing order. Higher priority processed first.
        VERY_HIGH > HIGH > MEDIUM > LOW > VERY_LOW
...
    jobOperationStatus:
      type: string
      enum:
        - PENDING
        - ABORTED
        - PAUSED
      example: PENDING
      description: >
        User-controllable job statuses: PENDING (resume), ABORTED (cancel),
        PAUSED (suspend).

        System-managed: IN_PROGRESS, COMPLETED, FAILED, CREATED.
...
    userMetadata:
      type: object
      additionalProperties: true
      default: {}
      description: >-
        Application-specific custom data container that can store arbitrary
        client information without affecting core operations
...
    createJobPayload:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/jobName'
        data:
          $ref: '#/components/schemas/jobPayload'
        priority:
          allOf:
            - $ref: '#/components/schemas/priority'
            - default: MEDIUM
        userMetadata:
          $ref: '#/components/schemas/userMetadata'
        traceparent:
          $ref: '#/components/schemas/traceparent'
        tracestate:
          $ref: '#/components/schemas/tracestate'
      required:
        - data
        - name
      additionalProperties: false
      description: >
        Input payload for creating a new job in the system.

        Contains all required configuration for job execution, including
        processing mode,

        custom parameters, metadata.


        Tracing fields (traceparent, tracestate) are optional:

        - If traceparent is provided, user's trace context is used (tracestate
        defaults to null if not provided)

        - If traceparent is not provided, the system automatically injects both
        traceparent and tracestate 
          from the active OpenTelemetry context using propagation.inject() (tracestate may still be null if not available)
    job:
      type: object
      required:
        - id
        - name
        - data
        - userMetadata
        - traceparent
        - status
        - percentage
        - creationTime
        - updateTime
        - priority
      description: >
        Complete job information with status and metadata.


        Comprehensive job response model containing all job details including
        configuration,

        execution status, progress tracking, and associated metadata. This
        schema represents

        the complete state of a job in the system and is returned by job
        retrieval operations.


        Includes optional stage data when requested via query parameters,
        allowing clients

        to retrieve the complete job hierarchy in a single request.
      properties:
        id:
          $ref: '#/components/schemas/jobId'
        status:
          $ref: '#/components/schemas/jobOperationStatusResponse'
        percentage:
          $ref: '#/components/schemas/percentage'
        creationTime:
          $ref: '#/components/schemas/creationTime'
        updateTime:
          $ref: '#/components/schemas/updateTime'
        name:
          $ref: '#/components/schemas/jobName'
        data:
          $ref: '#/components/schemas/jobPayload'
        priority:
          $ref: '#/components/schemas/priority'
        userMetadata:
          $ref: '#/components/schemas/userMetadata'
        traceparent:
          $ref: '#/components/schemas/traceparent'
        tracestate:
          $ref: '#/components/schemas/tracestate'
        stages:
          type: array
          items:
            $ref: '#/components/schemas/stageResponse'
...
    defaultOkMessage:
      type: object
      description: |
        Standard success response structure used for operations that don't
        return entity data, providing a standardized confirmation message.
      properties:
        code:
          $ref: '#/components/schemas/successMessages'
      required:
        - code
    baseErrorResponse:
      type: object
      description: >
        Foundation schema for standardized error responses across all API
        endpoints.

        Provides consistent structure with human-readable message and
        common-readable code fields

        that are extended by individual endpoints to define their specific error
        scenarios.
      properties:
        message:
          type: string
          description: Human-readable error message describing the issue
        code:
          type: string
          description: common-readable error code indicating the specific error type
      required:
        - message
        - code
    internalErrorsResponse:
      allOf:
        - $ref: '#/components/schemas/baseErrorResponse'
        - type: object
          description: >
            Provides a unified error response format for internal server errors,
            including a common-readable error code and a clear, human-readable
            message to help clients identify and handle issues consistently.
          properties:
            code:
              enum:
                - DATABASE_RELATED_ERROR
                - UNKNOWN_ERROR
          required:
            - code
    badRequestResponse:
      allOf:
        - $ref: '#/components/schemas/baseErrorResponse'
        - type: object
          description: >
            Standardized error response structure for bad request errors,
            providing a consistent

            format for common-readable error codes and human-readable messages.
          properties:
            code:
              enum:
                - VALIDATION_ERROR
          required:
            - code
    jobNotFoundResponse:
      allOf:
        - $ref: '#/components/schemas/baseErrorResponse'
        - type: object
          description: >
            Standardized error response for cases where a requested job is not
            found in the system.

            Provides a consistent format for common-readable error codes and
            human-readable messages.
          properties:
            code:
              enum:
                - JOB_NOT_FOUND
          required:
            - code
...